requirements.txt
python-dateutil >= 2.6.0
singer-python >= 2.1.4
singer-tools >= 0.4.0
requests >= 2.18.3
mailchimp3 >= 2.0.15
